requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0
azure-storage-blob==12.20.0
python-dotenv==1.0.0
Flask==3.0.0
flask-cors==4.0.0
